requirements.txt
google-generativeai
python-dotenv
datasets
requests
torch
transformers
